+++ src/select-input/select-input.tsx
@@ -1,13 +1,14 @@
 import {
   computed, defineComponent, ref, SetupContext, toRefs,
 } from '@vue/composition-api';
-import Popup from '../popup';
+import Popup, { PopupVisibleChangeContext } from '../popup';
 import props from './props';
 import { TdSelectInputProps } from './type';
 import useSingle from './useSingle';
 import useMultiple from './useMultiple';
 import useOverlayInnerStyle from './useOverlayInnerStyle';
 import { useConfig } from '../config-provider/useConfig';
+import { renderTNodeJSX } from '../utils/render-tnode';
 
 export default defineComponent({
   name: 'TSelectInput',
@@ -27,9 +28,22 @@ export default defineComponent({
       multiple, value, popupVisible, borderless,
     } = toRefs(props);
 
-    const { commonInputProps, onInnerClear, renderSelectSingle } = useSingle(props, context);
-    const { renderSelectMultiple } = useMultiple(props, context);
-    const { tOverlayInnerStyle, innerPopupVisible, onInnerPopupVisibleChange } = useOverlayInnerStyle(props);
+    const {
+      commonInputProps, singleInputValue, onInnerClear, renderSelectSingle,
+    } = useSingle(props, context);
+    const { multipleInputValue, renderSelectMultiple } = useMultiple(props, context);
+    const { tOverlayInnerStyle, innerPopupVisible, onInnerPopupVisibleChange } = useOverlayInnerStyle(props, {
+      afterHidePopup: onInnerBlur,
+    });
+
+    // SelectInput.blur is not equal to Input or TagInput, example: click popup panel.
+    // if trigger blur on click popup panel, filter data of tree select can not be checked.
+    function onInnerBlur(ctx: PopupVisibleChangeContext) {
+      const inputValue = props.multiple ? multipleInputValue.value : singleInputValue.value;
+      const params: Parameters<TdSelectInputProps['onBlur']>[1] = { e: ctx.e, inputValue };
+      props.onBlur?.(props.value, params);
+      context.emit('blur', props.value, params);
+    }
 
     const classes = computed(() => [
       `${classPrefix.value}-select-input`,
@@ -89,12 +103,13 @@ export default defineComponent({
       </Popup>
     );
 
+    const tipsNode = renderTNodeJSX(this, 'tips');
     return (
       <div ref="selectInputRef" class={this.classes}>
         {mainContent}
-        {this.tips && (
+        {tipsNode && (
           <div class={`${this.classPrefix}-input__tips ${this.classPrefix}-input__tips--${this.status || 'normal'}`}>
-            {this.tips}
+            {tipsNode}
           </div>
         )}
       </div>