+++ src/select-input/useMultiple.tsx
@@ -1,12 +1,12 @@
 import {
-  SetupContext, computed, ref, toRefs, getCurrentInstance,
+  SetupContext, computed, ref, toRefs,
 } from '@vue/composition-api';
 import isObject from 'lodash/isObject';
 import Vue from 'vue';
-import { TdSelectInputProps, SelectInputChangeContext, SelectInputKeys } from './type';
+import { TdSelectInputProps, SelectInputKeys } from './type';
 import { SelectInputCommonProperties } from './interface';
 import { InputValue } from '../input';
-import TagInput, { TagInputValue, InputValueChangeContext } from '../tag-input';
+import TagInput, { TagInputValue, InputValueChangeContext, TagInputProps } from '../tag-input';
 import Loading from '../loading';
 import useDefaultValue from '../hooks/useDefaultValue';
 import { usePrefixClass } from '../hooks/useConfig';
@@ -17,16 +17,15 @@ export interface RenderSelectMultipleParams {
   popupVisible: boolean;
 }
 
-const DEFAULT_KEYS = {
+const DEFAULT_KEYS: SelectInputKeys = {
   label: 'label',
-  key: 'key',
+  value: 'value',
   children: 'children',
 };
 
 export default function useMultiple(props: TdSelectInputProps, context: SetupContext) {
   const { inputValue } = toRefs(props);
   const classPrefix = usePrefixClass();
-  const instance = getCurrentInstance();
   const tagInputRef = ref();
   const [tInputValue, setTInputValue] = useDefaultValue(
     inputValue,
@@ -45,19 +44,17 @@ export default function useMultiple(props: TdSelectInputProps, context: SetupCon
 
   const tPlaceholder = computed<string>(() => (!tags.value || !tags.value.length ? props.placeholder : ''));
 
-  const onTagInputChange = (val: TagInputValue, context: SelectInputChangeContext) => {
+  const onTagInputChange: TagInputProps['onChange'] = (val, ctx) => {
     // 避免触发浮层的显示或隐藏
-    if (context.trigger === 'tag-remove') {
-      context.e?.stopPropagation();
+    if (ctx.trigger === 'tag-remove') {
+      ctx.e?.stopPropagation();
     }
-    props.onTagChange?.(val, context);
-    instance.emit('tag-change', val, context);
+    props.onTagChange?.(val, ctx);
+    context.emit('tag-change', val, ctx);
   };
-
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   const renderSelectMultiple = (p: RenderSelectMultipleParams, h: Vue.CreateElement) => {
     const tagInputProps = {
-      ...props.tagInputProps,
       ...p.commonInputProps,
       tagProps: props.tagProps,
       label: props.label,
@@ -78,28 +75,38 @@ export default function useMultiple(props: TdSelectInputProps, context: SetupCon
         ...props.inputProps,
       },
       suffixIcon: !props.disabled && props.loading ? () => <Loading loading size="small" /> : props.suffixIcon,
+      ...props.tagInputProps,
     };
+
     return (
       <TagInput
         ref="tagInputRef"
         scopedSlots={context.slots}
         props={tagInputProps}
         on={{
-          'input-change': (val: InputValue, context: InputValueChangeContext) => {
-            // 筛选器统一特性：筛选器按下回车时不清空输入框
-            if (context?.trigger === 'enter') return;
-            setTInputValue(val, { trigger: context.trigger, e: context.e });
+          'input-change': (val: InputValue, ctx: InputValueChangeContext) => {
+            /**
+             * 筛选器统一特性：
+             * 1. 筛选器按下回车时不清空输入框;
+             * 2. SelectInput 的失焦不等于 TagInput。如点击下拉面板时，TagInput 失去焦点，但 SelectInput 依旧保持聚焦，允许继续选择。
+             */
+            if (ctx.trigger === 'enter' || ctx.trigger === 'blur') return;
+            setTInputValue(val, { trigger: ctx.trigger, e: ctx.e });
           },
+          ...context.listeners,
         }}
         onChange={onTagInputChange}
         onClear={p.onInnerClear}
-        onBlur={(val: TagInputValue, context: { inputValue: InputValue; e: FocusEvent }) => {
-          props.onBlur?.(props.value, { ...context, tagInputValue: val });
-          instance.emit('blur', props.value, { ...context, tagInputValue: val });
+        // [Important Info]: SelectInput.blur is not equal to TagInput, example: click popup panel
+        onFocus={(val: TagInputValue, ctx: { inputValue: InputValue; e: FocusEvent }) => {
+          const params = { ...ctx, tagInputValue: val };
+          props.onFocus?.(props.value, params);
+          context.emit('focus', props.value, params);
         }}
-        onFocus={(val: TagInputValue, context: { inputValue: InputValue; e: FocusEvent }) => {
-          props.onFocus?.(props.value, { ...context, tagInputValue: val });
-          instance.emit('focus', props.value, { ...context, tagInputValue: val });
+        onEnter={(val: TagInputValue, ctx: { e: KeyboardEvent; inputValue: InputValue }) => {
+          const params = { ...ctx, tagInputValue: val };
+          props.onEnter?.(props.value, params);
+          context.emit('focus', props.value, params);
         }}
       />
     );
@@ -109,6 +116,7 @@ export default function useMultiple(props: TdSelectInputProps, context: SetupCon
     tags,
     tPlaceholder,
     tagInputRef,
+    multipleInputValue: tInputValue,
     renderSelectMultiple,
   };
 }