+++ test/utils/events.js
@@ -1,0 +1,23 @@
+export const EVENTS_MAP = {
+  mousedown: 'mouseDown',
+  mouseenter: 'mouseEnter',
+  mouseleave: 'mouseLeave',
+  mousemove: 'mouseMove',
+  mouseout: 'mouseOut',
+  mouseover: 'mouseOver',
+  mouseup: 'mouseUp',
+  keydown: 'keyDown',
+  keypress: 'keyPress',
+  keyup: 'keyUp',
+  dragend: 'dragEnd',
+  dragenter: 'dragEnter',
+  dragexit: 'dragExit',
+  dragleave: 'dragLeave',
+  dragover: 'dragOver',
+  dragstart: 'dragStart',
+  compositionend: 'compositionEnd',
+  compositionstart: 'compositionStart',
+  compositionupdate: 'compositionUpdate',
+};
+
+export default EVENTS_MAP;